commandresponse: Reorganize classes.
This structure adds a iterator module to store all iterators separate from the
command responses.
