+++ JMPDComm/src/main/java/com/anpmech/mpd/commandresponse/GenreResponse.java
@@ -28,11 +28,10 @@
 package com.anpmech.mpd.commandresponse;
 
 import com.anpmech.mpd.ResponseObject;
+import com.anpmech.mpd.commandresponse.iterator.GenreIterator;
 import com.anpmech.mpd.connection.CommandResult;
 import com.anpmech.mpd.item.Genre;
-import com.anpmech.mpd.item.Music;
 
-import java.util.Iterator;
 import java.util.ListIterator;
 
 /**
@@ -41,11 +40,6 @@ import java.util.ListIterator;
  * <p>This class is immutable, thus, thread-safe.</p>
  */
 public class GenreResponse extends ObjectResponse<Genre> {
-
-    /**
-     * This is the beginning block token to find for this multi-line response.
-     */
-    private static final String[] BLOCK_TOKEN = {Music.RESPONSE_GENRE};
 
     /**
      * The class log identifier.
@@ -101,47 +95,10 @@ public class GenreResponse extends ObjectResponse<Genre> {
      * Returns a count of how many objects this {@code Collection} contains.
      *
      * @return how many objects this {@code Collection} contains, or {@link Integer#MAX_VALUE}
-     * if there are more than {@link Integer#MAX_VALUE} elements in this
-     * {@code Collection}.
+     * if there are more than {@link Integer#MAX_VALUE} elements in this {@code Collection}.
      */
     @Override
     public int size() {
-        return CommandResponse.SingleLineResultIterator.size(mResult, BLOCK_TOKEN);
-    }
-
-    /**
-     * This class instantiates an {@link Iterator} to iterate over {@link Genre} entries.
-     */
-    private static final class GenreIterator extends
-            CommandResponse.SingleLineResultIterator<Genre> {
-
-        /**
-         * The class log identifier.
-         */
-        private static final String TAG = "GenreIterator";
-
-        /**
-         * Sole constructor.
-         *
-         * @param result   The MPD protocol command result.
-         * @param position The position relative to the result to initiate the {@link #mPosition}
-         *                 to.
-         * @throws IllegalArgumentException if the position parameter is less than 0.
-         */
-        private GenreIterator(final String result, final int position) {
-            super(result, position, BLOCK_TOKEN);
-        }
-
-        /**
-         * This method instantiates the {@link Genre} object with a block from the MPD server
-         * response.
-         *
-         * @param responseBlock The MPD server response to instantiate the Genre entry with.
-         * @return The Genre entry.
-         */
-        @Override
-        Genre instantiate(final String responseBlock) {
-            return new Genre(responseBlock);
-        }
+        return GenreIterator.size(mResult);
     }
 }
