+++ JMPDComm/src/main/java/com/anpmech/mpd/commandresponse/PlaylistFileResponse.java
@@ -28,14 +28,11 @@
 package com.anpmech.mpd.commandresponse;
 
 import com.anpmech.mpd.ResponseObject;
-import com.anpmech.mpd.Tools;
+import com.anpmech.mpd.commandresponse.iterator.PlaylistFileIterator;
 import com.anpmech.mpd.connection.CommandResult;
 import com.anpmech.mpd.item.PlaylistFile;
-import com.anpmech.mpd.item.Stream;
 
-import java.util.Iterator;
 import java.util.ListIterator;
-import java.util.NoSuchElementException;
 
 /**
  * This class contains methods used to process {@link PlaylistFile} entries from a MPD response.
@@ -43,11 +40,6 @@ import java.util.NoSuchElementException;
  * <p>This class is immutable, thus, thread-safe.</p>
  */
 public class PlaylistFileResponse extends ObjectResponse<PlaylistFile> {
-
-    /**
-     * This is the beginning block token to find for this multi-line response.
-     */
-    private static final String[] BLOCK_TOKEN = {PlaylistFile.RESPONSE_PLAYLIST_FILE};
 
     /**
      * The class log identifier.
@@ -103,161 +95,10 @@ public class PlaylistFileResponse extends ObjectResponse<PlaylistFile> {
      * Returns a count of how many objects this {@code Collection} contains.
      *
      * @return how many objects this {@code Collection} contains, or {@link Integer#MAX_VALUE}
-     * if there are more than {@link Integer#MAX_VALUE} elements in this
-     * {@code Collection}.
+     * if there are more than {@link Integer#MAX_VALUE} elements in this {@code Collection}.
      */
     @Override
     public int size() {
-        final Iterator<Void> iterator = new NoopIterator(mResult);
-
-        return CommandResponse.AbstractObjectResultIterator.count(iterator);
-    }
-
-    /**
-     * This class instantiates an {@link Iterator} to iterate over {@link PlaylistFile} entries.
-     */
-    private abstract static class AbstractPlaylistFileIterator<T> extends
-            CommandResponse.SingleLineResultIterator<T> {
-
-        /**
-         * The class log identifier.
-         */
-        private static final String TAG = "PlaylistFileIterator";
-
-        /**
-         * Sole constructor.
-         *
-         * @param result   The MPD protocol command result.
-         * @param position The position relative to the result to initiate the {@link #mPosition}
-         *                 to.
-         * @throws IllegalArgumentException if the position parameter is less than 0.
-         */
-        private AbstractPlaylistFileIterator(final String result, final int position) {
-            super(result, position, BLOCK_TOKEN);
-        }
-
-        /**
-         * This method checks to see at the key's value at {@code position} is
-         * {@link Stream#PLAYLIST_NAME}.
-         *
-         * @param position The position of the key of the value to check.
-         * @return True if the value of the key at {@code position} matches
-         * {@link Stream#PLAYLIST_NAME}, false otherwise.
-         */
-        private boolean isStreamValue(final int position) {
-            int start = position;
-
-            if (start != -1) {
-                start += PlaylistFile.RESPONSE_PLAYLIST_FILE.length() + 2;
-            }
-
-            return mResult.regionMatches(start, Stream.PLAYLIST_NAME, 0,
-                    Stream.PLAYLIST_NAME.length());
-        }
-
-        /**
-         * This method returns the index of the next beginning token in relation to the current
-         * position.
-         *
-         * @return The next beginning token in relation to the current position.
-         */
-        @Override
-        protected int nextIndexBegin() {
-            int index = super.nextIndexBegin();
-
-            if (isStreamValue(index)) {
-                index = Tools.getNextKeyIndex(mResult, index + 1, mBeginBlockTokens);
-            }
-
-            return index;
-        }
-
-        /**
-         * This method returns the index of the prior beginning token in relation to the current
-         * position.
-         *
-         * @return The prior beginning token in relation to the current position.
-         */
-        @Override
-        protected int previousIndexBegin() {
-            int index = super.previousIndexBegin();
-
-            if (isStreamValue(index)) {
-                index = previousIndexBegin(index);
-            }
-
-            return index;
-        }
-    }
-
-    /**
-     * This class implements a {@link AbstractPlaylistFileIterator} simply for counting
-     * iterations with less required garbage collection.
-     */
-    private static final class NoopIterator extends AbstractPlaylistFileIterator<Void> {
-
-        /**
-         * Sole constructor.
-         *
-         * @param result The MPD protocol command result.
-         */
-        private NoopIterator(final String result) {
-            super(result, 0);
-        }
-
-        /**
-         * This method instantiates the {@link PlaylistFile} object with a block from the MPD
-         * server response.
-         *
-         * @param responseBlock The MPD server response to instantiate the Music entry with.
-         * @return The PlaylistFile entry.
-         */
-        @Override
-        Void instantiate(final String responseBlock) {
-            return null;
-        }
-
-        /**
-         * Returns the next object in the iteration.
-         *
-         * @return the next object.
-         * @throws NoSuchElementException If there are no more elements.
-         * @see #hasNext
-         */
-        @Override
-        public Void next() {
-            return voidNext();
-        }
-    }
-
-    /**
-     * This class instantiates an {@link Iterator} to iterate over {@link PlaylistFile} entries.
-     */
-    private static final class PlaylistFileIterator
-            extends AbstractPlaylistFileIterator<PlaylistFile> {
-
-        /**
-         * Sole constructor.
-         *
-         * @param result   The MPD protocol command result.
-         * @param position The position relative to the result to initiate the {@link #mPosition}
-         *                 to.
-         * @throws IllegalArgumentException if the position parameter is less than 0.
-         */
-        private PlaylistFileIterator(final String result, final int position) {
-            super(result, position);
-        }
-
-        /**
-         * This method instantiates the {@link PlaylistFile} object with a block from the MPD
-         * server response.
-         *
-         * @param responseBlock The MPD server response to instantiate the Music entry with.
-         * @return The PlaylistFile entry.
-         */
-        @Override
-        PlaylistFile instantiate(final String responseBlock) {
-            return PlaylistFile.byResponse(responseBlock);
-        }
+        return PlaylistFileIterator.size(mResult);
     }
 }
