+++ JMPDComm/src/main/java/com/anpmech/mpd/commandresponse/iterator/NoopPlaylistFileIterator.java
@@ -25,8 +25,48 @@
  * THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
+package com.anpmech.mpd.commandresponse.iterator;
+
+import com.anpmech.mpd.item.PlaylistFile;
+
+import java.util.NoSuchElementException;
+
 /**
- * This package contains {@link java.util.Collection}s for processing {@link
- * com.anpmech.mpd.connection.CommandResult} Objects.
+ * This class implements a {@link AbstractPlaylistFileIterator} simply for counting
+ * iterations with less required garbage collection.
  */
-package com.anpmech.mpd.commandresponse;
+class NoopPlaylistFileIterator extends AbstractPlaylistFileIterator<Void> {
+
+    /**
+     * Sole constructor.
+     *
+     * @param result The MPD protocol command result.
+     */
+    NoopPlaylistFileIterator(final String result) {
+        super(result, 0);
+    }
+
+    /**
+     * This method instantiates the {@link PlaylistFile} object with a block from the MPD
+     * server response.
+     *
+     * @param responseBlock The MPD server response to instantiate the Music entry with.
+     * @return The PlaylistFile entry.
+     */
+    @Override
+    protected Void instantiate(final String responseBlock) {
+        return null;
+    }
+
+    /**
+     * Returns the next object in the iteration.
+     *
+     * @return the next object.
+     * @throws NoSuchElementException If there are no more elements.
+     * @see #hasNext
+     */
+    @Override
+    public Void next() {
+        return voidNext();
+    }
+}
